# perf(cache): improve drafts cache for perf